audio: remove mixeng specific code from backends
Backends no longer have to deal with mixeng, they just receive a buffer
in the correct sample format, all mixeng logic is now in the audio.c
(and mixeng.c).  Backends also do not have to deal with soft voices.

Backends now have two way to read/write sound:
* write and read functions: similar to old read/write functions, except
  that they actually read/write the data to the backend instead of
  placing it into the mixeng buffer.  You no longer need run_in/run_out
  afterwards.
* get_buffer_out/put_buffer_out: the first function returns a buffer
  that can hold some audio data.  The caller fills this buffer (maybe
  partially) and calls put_buffer to actually write the data.  This way
  we can save copying the buffer in some cases (for example mmaped
  audio).  Similarly there's get_buffer_in/put_buffer_in for reading.

Backends only have to support one access method, but they can support
both if they have efficient implementation for both cases.

Please note that the read/write functions are not used directly in this
commit yet.

Signed-off-by: Kővágó, Zoltán <DirtY.iCE.hu@gmail.com>

---

There's one thing I'm not sure about: backends used to return in
hw->samples a buffer size, and mixeng allocated that large buffer for
itself.  The old code pretty much required that the mixeng buffer and
the destination buffer have the same size, but it's no longer required.
So maybe it would be a better idea to drop buffer_size_* functions (the
replacement of old hw->samples) and specify it as a (global) -audiodev
option.

Also please double check that coreaudio backend works.  I do not have a
mac to test.